requirements.txt
streamlit>=1.24.0
python-docx>=0.8.11
pdfkit>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.2.0
spacy>=3.6.0
pandas>=2.0.0
pypdf>=3.12.0
docx2txt>=0.8